[AUTO Plugin] Rename Multi plugin to ov_multi_plugin and move the src files to new place.
