+++ inference-engine/src/CMakeLists.txt
@@ -20,7 +20,3 @@ endif()
 if(ENABLE_GNA)
     add_subdirectory(gna_plugin)
 endif()
-
-if(ENABLE_MULTI)
-    add_subdirectory(multi_device)
-endif()
