+++ inference-engine/tests/functional/plugin/conformance/test_runner/api_conformance_runner/include/api_conformance_helpers.hpp
@@ -19,12 +19,12 @@ inline const std::string getPluginLibNameByDevice(const std::string& deviceName)
             { "AUTO", "AutoPlugin" },
             { "HDDL", "HDDLPlugin" },
             { "VPUX", "VPUXPlugin" },
-            { "AUTO", "MultiDevicePlugin" },
+            { "AUTO", "ov_auto_plugin" },
             { "CPU", "MKLDNNPlugin" },
             { "GNA", "GNAPlugin" },
             { "GPU", "ov_intel_gpu_plugin" },
             { "HETERO", "ov_hetero_plugin" },
-            { "MULTI", "MultiDevicePlugin" },
+            { "MULTI", "ov_multi_plugin" },
             { "MYRIAD", "myriadPlugin" },
             { "TEMPLATE", "templatePlugin" },
     };
