+++ inference-engine/tests/functional/plugin/shared/CMakeLists.txt
@@ -12,8 +12,8 @@ if(ENABLE_HETERO)
     list(APPEND DEPENDENCIES ov_hetero_plugin)
 endif()
 
-if(ENABLE_MULTI)
-    list(APPEND DEPENDENCIES MultiDevicePlugin)
+if(ENABLE_AUTO OR ENABLE_MULTI)
+    list(APPEND DEPENDENCIES ov_auto_plugin)
 endif()
 
 # remove once CVS-69781 is fixed
