+++ inference-engine/tests/unit/auto/auto_select_device_failed_test.cpp
@@ -15,7 +15,7 @@
 #include <ngraph_functions/subgraph_builders.hpp>
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
-#include "plugin/mock_multi_device_plugin.hpp"
+#include "plugin/mock_auto_device_plugin.hpp"
 #include "cpp/ie_plugin.hpp"
 
 using ::testing::MatcherCast;
