+++ src/plugins/CMakeLists.txt
@@ -9,3 +9,7 @@ endif()
 if(ENABLE_INTEL_GPU)
     add_subdirectory(intel_gpu)
 endif()
+
+if(ENABLE_AUTO OR ENABLE_MULTI)
+    add_subdirectory(auto)
+endif()
