+++ src/plugins/auto/executable_network.cpp
@@ -14,15 +14,15 @@
 #include "ie_icore.hpp"
 #include "ie_metric_helpers.hpp"
 #include <ie_plugin_config.hpp>
-#include "multi_device_exec_network.hpp"
-#include "multi_device_async_infer_request.hpp"
-#include "multi_device_plugin.hpp"
+#include "executable_network.hpp"
+#include "async_infer_request.hpp"
+#include "plugin.hpp"
 
 #include "ngraph/opsets/opset1.hpp"
 #include "transformations/utils/utils.hpp"
 #include "utils/log_util.hpp"
 
-#include "multi_itt.hpp"
+#include "itt.hpp"
 // ------------------------------MultiDeviceExecutableNetwork----------------------------
 namespace MultiDevicePlugin {
 using namespace InferenceEngine;
