+++ src/plugins/auto/infer_request.hpp
