+++ src/plugins/auto/utils/non_copyable.hpp
