+++ src/plugins/auto/utils/singleton.hpp
