+++ tools/deployment_manager/configs/darwin.json
@@ -21,7 +21,7 @@
         "runtime/lib/intel64/libinference_engine_preproc.so",
         "runtime/lib/intel64/libinference_engine_c_api.dylib",
         "runtime/lib/intel64/libov_hetero_plugin.so",
-        "runtime/lib/intel64/libMultiDevicePlugin.so",
+        "runtime/lib/intel64/libov_auto_plugin.so",
         "runtime/lib/intel64/libngraph.dylib",
         "runtime/lib/intel64/libfrontend_common.dylib",
         "runtime/lib/intel64/libir_ov_frontend.dylib",
