+++ tools/deployment_manager/configs/linux.json
@@ -27,7 +27,7 @@
         "runtime/lib/intel64/libinference_engine_preproc.so",
         "runtime/lib/intel64/libinference_engine_c_api.so",
         "runtime/lib/intel64/libov_hetero_plugin.so",
-        "runtime/lib/intel64/libMultiDevicePlugin.so",
+        "runtime/lib/intel64/libov_auto_plugin.so",
         "runtime/lib/intel64/libngraph.so",
         "runtime/lib/intel64/libfrontend_common.so",
         "runtime/lib/intel64/libir_ov_frontend.so",
